refactor(nimbus-provider): streamline color mode and router types integration
- Removed outdated router-types file and consolidated type definitions into nimbus-provider.types.ts for better organization.
- Updated ColorModeProvider to import types directly from the new types file, enhancing clarity and maintainability.
- Simplified imports in nimbus-provider.tsx by removing unnecessary props and ensuring type safety across components.
